Implement @piwakk's suggestions + fix bugs
