+++ frontend/src/components/associations/marketplace/History.tsx
@@ -7,6 +7,7 @@ import Card from "react-bootstrap/Card";
 import { Pagination } from "../../utils/Pagination";
 import { formatDate } from "../../../utils/format";
 import { TransactionStatus } from "../../../models/associations/marketplace";
+import { Table } from "../../utils/table/Table";
 
 export const AssociationMarketplaceHistory = ({ association }) => {
     const user = useContext(UserContext);
@@ -14,7 +15,7 @@ export const AssociationMarketplaceHistory = ({ association }) => {
 
     return (
         <>
-            <div className={"float-right"}>
+            <div className={"float-center"}>
                 <a
                     href={"/associations/" + marketplaceId + "/marketplace"}
                     className={"btn btn-primary"}
@@ -28,46 +29,94 @@ export const AssociationMarketplaceHistory = ({ association }) => {
     );
 };
 
-// Transactions informations
-const TransactionHistory = ({ marketplaceId, user }) => (
-    <Pagination
-        apiKey={[
-            "marketplace.transactions.list",
-            marketplaceId,
-            { page_size: 10, buyer: user?.id },
-        ]}
-        apiMethod={api.transactions.list}
-        render={(transactions, paginationControl) => (
-            <Container>
-                <PageTitle>Mes commandes</PageTitle>
+const TransactionHistory = ({ marketplaceId, user }) => {
+    const columns = [
+        {
+            key: "product",
+            header: "Produit",
+            render: (transaction) => (
+                <>
+                    <strong>{transaction.product.name}</strong>{" "}
+                    <span className="text-muted">x{transaction.quantity}</span>
+                </>
+            ),
+        },
+        {
+            key: "value",
+            header: "Valeur",
+            render: (transaction) => <strong>{transaction.value} €</strong>,
+            headerClassName: "text-center",
+            cellClassName: "text-center",
+        },
+        {
+            key: "date",
+            header: "Date",
+            render: (transaction) => formatDate(new Date(transaction.date)),
+            headerClassName: "text-center",
+            cellClassName: "text-center",
+        },
+        {
+            key: "status",
+            header: "Status",
+            render: (transaction) => {
+                let status = {};
 
-                {transactions.length !== 0 ? (
-                    <>
-                        <TransactionsTable transactions={transactions} />
-                        {paginationControl}
-                    </>
-                ) : (
-                    <NoTransactionMessage />
-                )}
-            </Container>
-        )}
-    />
-);
+                status[TransactionStatus.Ordered] = (
+                    <span className="tag tag-blue">Commandée</span>
+                );
 
-const TransactionsTable = ({ transactions }) => {
+                status[TransactionStatus.Validated] = (
+                    <span className="tag tag-lime">Validée</span>
+                );
+
+                status[TransactionStatus.Delivered] = (
+                    <span className="tag tag-green">Délivrée</span>
+                );
+
+                status[TransactionStatus.Cancelled] = (
+                    <span className="tag tag-red">Annulée</span>
+                );
+
+                status[TransactionStatus.Refunded] = (
+                    <span className="tag tag-yellow">Remboursée</span>
+                );
+
+                status[TransactionStatus.Rejected] = (
+                    <span className="tag tag-red">Refusée</span>
+                );
+
+                return status[transaction.status];
+            },
+            headerClassName: "text-center",
+            cellClassName: "text-center",
+        },
+    ];
+
     return (
-        <Card>
-            <table className="table card-table table-vcenter">
-                <tbody>
-                    {transactions.map((transaction, index) => (
-                        <PurchaseTransactionLine
-                            key={index}
-                            transaction={transaction}
-                        />
-                    ))}
-                </tbody>
-            </table>
-        </Card>
+        <Pagination
+            apiKey={[
+                "marketplace.transactions.list",
+                marketplaceId,
+                { page_size: 10, buyer: user?.id },
+            ]}
+            apiMethod={api.transactions.list}
+            render={(transactions, paginationControl) => (
+                <Container>
+                    <PageTitle>Mes commandes</PageTitle>
+
+                    {transactions.length !== 0 ? (
+                        <>
+                            <Card>
+                                <Table columns={columns} data={transactions} />
+                            </Card>
+                            {paginationControl}
+                        </>
+                    ) : (
+                        <NoTransactionMessage />
+                    )}
+                </Container>
+            )}
+        />
     );
 };
 
@@ -81,47 +130,41 @@ const NoTransactionMessage = () => (
     </Card>
 );
 
-const PurchaseTransactionLine = ({ transaction }) => {
-    let status = <p />;
-
-    if (transaction.status === TransactionStatus.Ordered) {
-        status = <span className="tag tag-blue">Commandée</span>;
-    }
-    if (transaction.status === TransactionStatus.Validated) {
-        status = <span className="tag tag-lime">Validée</span>;
-    }
-    if (transaction.status === TransactionStatus.Delivered) {
-        status = <span className="tag tag-green">Délivrée</span>;
-    }
-    if (transaction.status === TransactionStatus.Cancelled) {
-        status = <span className="tag tag-red">Annulée</span>;
-    }
-    if (transaction.status === TransactionStatus.Refunded) {
-        status = <span className="tag tag-yellow">Remboursée</span>;
-    }
-    if (transaction.status === TransactionStatus.Rejected) {
-        status = <span className="tag tag-red">Refusée</span>;
-    }
-
-    return (
-        <tr>
-            <td>
-                <strong>{transaction.product.name}</strong>{" "}
-                <span className="text-muted">x{transaction.quantity}</span>
-            </td>
-            <td className="text-center">
-                <strong>{transaction.value}€</strong>
-            </td>
-            <td className="text-right">
-                {formatDate(new Date(transaction.date))}
-            </td>
-            <td className="text-right">{status}</td>
-        </tr>
-    );
-};
-
-// Funding informations
 const FundingHistory = ({ marketplaceId, user }) => {
+    const columns = [
+        {
+            key: "label",
+            header: "",
+            render: (funding) => "Argent versé sur le compte :",
+        },
+        {
+            key: "value",
+            header: "Valeur",
+            render: (funding) => funding.value + " €",
+            headerClassName: "text-center",
+            cellClassName: "text-center",
+        },
+        {
+            key: "date",
+            header: "Date",
+            render: (funding) => formatDate(new Date(funding.date)),
+            headerClassName: "text-center",
+            cellClassName: "text-center",
+        },
+        {
+            key: "value",
+            header: "Valeur",
+            render: (funding) =>
+                funding.status === "FUNDED" ? (
+                    <span className="tag tag-blue">Versé</span>
+                ) : funding.status === "REFUNED" ? (
+                    <span className="tag tag-yellow">Remboursé</span>
+                ) : null,
+            headerClassName: "text-center",
+            cellClassName: "text-center",
+        },
+    ];
+
     return (
         <Pagination
             apiKey={[
@@ -136,7 +179,9 @@ const FundingHistory = ({ marketplaceId, user }) => {
 
                     {fundings.length !== 0 ? (
                         <>
-                            <FundingTable fundings={fundings} />
+                            <Card>
+                                <Table columns={columns} data={fundings} />
+                            </Card>
                             {paginationControl}
                         </>
                     ) : (
@@ -145,20 +190,6 @@ const FundingHistory = ({ marketplaceId, user }) => {
                 </Container>
             )}
         />
-    );
-};
-
-const FundingTable = ({ fundings }) => {
-    return (
-        <Card>
-            <table className="table card-table table-vcenter">
-                <tbody>
-                    {fundings.map((funding, index) => (
-                        <FundingLine key={index} funding={funding} />
-                    ))}
-                </tbody>
-            </table>
-        </Card>
     );
 };
 
@@ -171,26 +202,3 @@ const NoFundingMessage = () => (
         </Card.Body>
     </Card>
 );
-
-const FundingLine = ({ funding }) => {
-    /* If the user put money  on their account */
-    let status = <p />;
-    if (funding.status === "FUNDED") {
-        status = <span className="tag tag-blue">Versé</span>;
-    } else if (funding.status === "REFUNED") {
-        status = <span className="tag tag-yellow">Remboursé</span>;
-    }
-
-    return (
-        <tr>
-            <td>Argent versé sur le compte :</td>
-            <td className={"text-center"}>
-                <strong>{funding.value}€</strong>
-            </td>
-            <td className="text-center">
-                {formatDate(new Date(funding.date))}
-            </td>
-            <td className="text-right">{status}</td>
-        </tr>
-    );
-};
